+++ task1_prac_tr1/models.py
@@ -13,7 +13,7 @@ Your app description
 
 
 class Constants(BaseConstants):
-    name_in_url = 'dynamic_portfolio_practice'
+    name_in_url = 'task1_prac_tr1'
     players_per_group = None
     num_rounds = 2
     initial_wealth_list = [120, 80]
